Fix backend to support decimal values for count-based tasks and fix quarterly task API endpoint
- Changed target_value from int to float in Pydantic schemas (TaskBase, TaskUpdate)
- Changed target_value column from Integer to Float in SQLAlchemy model
- Added migration 021 to convert database column from INTEGER to REAL
- Fixed quarterly task handler to pass year_start_date in URL path instead of request body
- Backend now accepts decimal values like 0.3 for count-based tasks

+++ backend/app/models/models.py
@@ -177,7 +177,7 @@ class Task(Base):
     # Task type and allocation
     task_type = Column(String(20), nullable=False, default='time')  # 'time', 'count', or 'boolean'
     allocated_minutes = Column(Integer, nullable=False)  # Time in minutes (for time-based tasks)
-    target_value = Column(Integer, nullable=True)  # Target for count-based tasks (e.g., 10 push-ups)
+    target_value = Column(Float, nullable=True)  # Target for count-based tasks (supports decimals like 0.3)
     unit = Column(String(50), nullable=True)  # Unit for count tasks (e.g., 'reps', 'glasses', 'miles')
     spent_minutes = Column(Integer, default=0)  # Actual time spent
